Move utils back to component
